Let's be loud. But in an intelligent way with decorum. They love that word. Wonder what it really means.
Been up all night researching. Had a great talk with a friend who was cynical at first but all you have to do is connect a few dots for the average critical thinking individual and it all starts to make sense. If you wanna throw in some well known conspiracy theories,  I'm fine with that. Hey I wonder if the FBI ever apologized to Coretta Scott King. Or the whole country for that matter. We have receipts on that one.  Found a letter addressed to me from 2023. Had my mom's handwriting all over it. It Had my social security on it. And several times she was supposed to talk to different people. I wonder why I wasn't involved in this.  It was MY mail? More stories like this to come if they will let me get it out. I'll leave you with this though. On the back of the letter, I was advised to get a lawyer and there were 4  firms printed right there with numbers to call. They specialized in civil rights cases.I never got the chance to call. It was hidden to me. Instead, wothout prior consent I was put into a program sponsored by our gov to care for me by surveillance 24/7.